Added support for more command line options and Fixed song playing error
Fixed a spelling mistake in a comment in Pipe.cpp
It now can check all command line arguments and added an argument named -subdir so that a user can specify the a specific folder within the music folder.
Fixed an error with getCurrentSong() failing in the play_next loop because of the playedFiles vector being cleared.